Home Page: Guest can see countries by region
Implements a presenter, which will display countries by their region
when the region is clicked. Creates a feature test for this. Also
places the countries in tables. This is done by passing the region id into params
and then passing that param into the presenter method, which looks up the countries by
that id.