Builds out decorator, world_bank_country, and indicators
Creates indicators method for world_bank_country which returns the indicators for the highest income for
world bank countries. Indicator class returns world_bank_indicators for highest_income and
also is structured to easily allow building more queries. Creates a country_decorator,
which takes an AR country and a WB country as arguments and allows those objects
to be accessed through the decorator.