Guest visits the country show page
Fixes the decorator to be a simple PORO and not a SimpleDelegator, which was
causing weird kernel errors and was also not necessary. Writes indicators method
for the decorator which returns a nested array of AR and WB objects to send to the view.
It definitely does some Ruby which might not be ideal but does do the thing. All tests passing
with 96% coverage. This completes major functionality for the project.